move function from kilo
